Update page_exec_test.go
upgrade

+++ example/tool/page_exec_test.go
@@ -11,7 +11,7 @@ func TestPageExec(t *testing.T) {
 
 	arr := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 
-	tool.PageExec(len(arr), 3, func(b, e, p int) (err error) {
+	tool.PageExec(int64(len(arr)), 3, func(b, e, p int) (err error) {
 		fmt.Println(fmt.Sprintf("%s\t:%v", t.Name(), arr[b:e]))
 		return
 	})
